renamed-pages-to-views-because-why-not?
